requirements_yolo.txt
# Requirements for YOLO-based egg detection
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Image processing
opencv-python-headless==4.8.1.78
numpy==1.24.4
pillow==10.0.0

# YOLO/PyTorch dependencies
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196

# Database and utilities
supabase==2.7.0
python-dotenv==1.0.0
